# app/api/v1/auth/auth.service.js
'use strict';

const appRequire = require('app-root-path').require;
const logger = appRequire('/config/logger/logger.config');
const bcrypt = require('bcryptjs');
const jwt = require('jsonwebtoken');
const HttpStatus = require('http-status-codes');
const AppStatus = require('./../common/app-status');
const _$ = require('./../common/constants');
const TeacherModel = require('./../teachers/teachers.model');
const ErrorMessage = require('../common/ErrorMessage');
const Result = require('../common/Result');
const AppError = require('./../common/AppError');

function handleLoginError() {
    const body = new ErrorMessage(
        AppStatus.LOGIN_NOT_VALID,
        AppStatus.getStatusText(AppStatus.LOGIN_NOT_VALID),
        null
    );

    return new Result(HttpStatus.UNAUTHORIZED, body);
}

function handleErrors(error) {
    switch (error.name) {
        case _$.LOGIN_ERROR:
            return handleLoginError();
        default:
            return AppError.handleGenericError();
    }
}

function throwLoginError(logMessage) {
    const error = new Error(logMessage);
    error.name = _$.LOGIN_ERROR;
    throw error;
}

const post = async function (request, jwtSecret) {
    let result = new Result();
    const username = request.username;
    const password = request.password;

    try {
        const teacher = await TeacherModel.findOne({
            username: username
        }).select('+password');
        if (!teacher) throwLoginError(AppStatus.getStatusText(AppStatus.TEACHER_NOT_FOUND));

        const isPasswordValid = await bcrypt.compare(password, teacher.password);
        if (!isPasswordValid) throwLoginError(AppStatus.getStatusText(AppStatus.PASSWORD_NOT_VALID));

        const token = await jwt.sign({id: teacher._id}, jwtSecret, {expiresIn: 7200}); //2 hours

        result.status = HttpStatus.OK;
        result.body = {
            auth: true,
            token: token
        };

    } catch (error) {
        logger.error(error);
        result = handleErrors(error);
    }

    return result;
};

module.exports = {
    post: post
};