# app/api/v1/spec/swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: Selnate Api
  description: This doc describes the apis and how to use them
basePath: /api/v1
...
schemes:
  - http
paths:
  /teachers:
    get:
      tags:
        - teachers
      description: Retrieve a list of teacher accounts.
      produces:
        - application/json
      responses:
        '200':
          description: Successful operation.
          schema:
            type: array
            items:
              $ref: '#/definitions/Teacher'
    post:
      tags:
        - teachers
      description: Create a new teacher account.
      parameters:
        - in: body
          name: body
          description: Teacher account information.
          required: true
          schema:
            $ref: '#/definitions/Teacher'
      responses:
        '200':
          description: Successful operation.
          schema:
            $ref: '#/definitions/Teacher'
        '400':
          description: Teacher not created.
          schema:
            $ref: '#/definitions/ErrorMessage'
  '/teachers/{id}':
    get:
      tags:
        - teachers
      description: Retrieve a teacher account based on id.
      parameters:
        - in: path
          name: id
          description: Teacher's id.
          required: true
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: Successful operation.
          schema:
            $ref: '#/definitions/Teacher'
        '404':
          description: Teacher not found.
          schema:
            $ref: '#/definitions/ErrorMessage'
    put:
      tags:
        - teachers
      description: Update a teacher's account based on id.
      parameters:
        - in: path
          name: id
          description: Teacher's id.
          required: true
          type: string
        - in: body
          name: body
          description: Teacher account.
          required: true
          schema:
            $ref: '#/definitions/Teacher'
      produces:
        - application/json
      responses:
        '200':
          description: Successful operation.
          schema:
            $ref: '#/definitions/Teacher'
        '400':
          description: Invalid teacher information.
          schema:
            $ref: '#/definitions/ErrorMessage'
        '404':
          description: Teacher not found.
          schema:
            $ref: '#/definitions/ErrorMessage'
    delete:
      tags:
        - teachers
      description: Deactivate a teacher account based on id.
      parameters:
        - in: path
          name: id
          description: Teacher's id.
          required: true
          type: string
      produces:
        - application/json
      responses:
        '204':
          description: Successful operation.
        '404':
          description: Teacher not found.
          schema:
            $ref: '#/definitions/ErrorMessage'
...
definitions:
  Teacher:
    type: object
    required:
      - firstName
      - lastName
      - username
      - password
    properties:
      id:
        type: integer
      firstName:
        type: string
      lastName:
        type: string
      username:
        type: string
      password:
        type: string
      isAdmin:
        type: boolean
        default: false
      isActive:
        type: boolean
        default: false
...
  ErrorMessage:
    type: object
    required:
      - code
      - message
      - detail
    properties:
      code:
        type: integer
      message:
        type: string
      detail:
        type: array
        items:
          type: string